Move CPU-related code into `cpu.rs` file
